requirements.txt
# Dipendenze Python per GrokMeme Bot

# Twitter API
tweepy>=4.14.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0
